refactor: useUsdcPriceHook & useTokens related hook (#309)
* fix: restructure useApproveCallbackhook

* fix: delete useApproveCallback

* fix: restructure useSwapCallbackhook

* fix: delete useSwapCallbackhook file

* fix: restructure useUSDCPricehook

* fix: restructure tokens hook

* fix: delete useUsdcHook file

* fix: some changes
