+++ src/state/pstake/hooks.ts
@@ -24,9 +24,11 @@ import { DAIe, PNG, USDC, USDCe, USDTe, axlUST } from 'src/constants/tokens';
 import { PairState, usePair, usePairs } from 'src/data/Reserves';
 import { usePairTotalSupplyHook } from 'src/data/multiChainsHooks';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
-import { useTokensHook, useUSDCPriceHook } from 'src/hooks/multiChainsHooks';
+import { useTokensHook } from 'src/hooks/tokens';
+import { useTokens } from 'src/hooks/tokens/evm';
 import usePrevious from 'src/hooks/usePrevious';
-import { useUSDCPrice } from 'src/hooks/useUSDCPrice';
+import { useUSDCPriceHook } from 'src/hooks/useUSDCPrice';
+import { useUSDCPrice } from 'src/hooks/useUSDCPrice/evm';
 import {
   updateMinichefStakingAllAprs,
   updateMinichefStakingAllData,
@@ -34,7 +36,6 @@ import {
 } from 'src/state/pstake/actions';
 import { usePairBalanceHook } from 'src/state/pwallet/multiChainsHooks';
 import { unwrappedToken } from 'src/utils/wrappedCurrency';
-import { useTokens } from '../../hooks/Tokens';
 import { useMiniChefContract, useRewardViaMultiplierContract, useStakingContract } from '../../hooks/useContract';
 import { tryParseAmount } from '../../state/pswap/hooks';
 import { AppState, useDispatch, useSelector } from '../index';
